Fix a compiler warning on gcc 6.1

+++ src/oscaroutput.cc
@@ -453,7 +453,8 @@ void OscarOutput<Format, Contents>::write_particledata(
                  data.pdgcode().string().c_str(), data.id(),
                  data.get_history().collisions_per_particle,
                  data.formation_time(), data.cross_section_scaling_factor(),
-                 data.get_history().id_process, data.get_history().process_type,
+                 data.get_history().id_process,
+                 static_cast<int>(data.get_history().process_type),
                  data.get_history().time_of_origin,
                  data.get_history().p1.string().c_str(),
                  data.get_history().p2.string().c_str());
